[first_exam] separation of source files and add package.json for using module import
